Add files via upload
“Password Manager” is a small-scale project used for storing your login 
information for all the websites you use and generating random passwords 
for users. The user will have the access to add, update and remove his details 
and he can fetch his data whenever needed.


It’s an GUI based application simple and easy to use. It has a lot’s of
feature which makes user experience better. In this application Passwords 
can be stored for various applications and details required for maintaining
the Passwords are application or website name, username, password and
email id. 

You would not get any problem while using it because it’s simple and
user friendly UI will make your work easy, faster and gives you a better 
result.